requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Templates
jinja2==3.1.2

# HTTP client
httpx==0.25.1

# Settings & Config
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Redis
redis==5.0.1

# LLM providers
openai==1.3.7
anthropic==0.7.7

# YouTube & Transcription
youtube-transcript-api==0.6.1
yt-dlp==2023.11.16
openai-whisper==20231117

# Article extraction
beautifulsoup4==4.12.2
readability-lxml==0.8.1

# Utilities
itsdangerous==2.1.2
